% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_analysis.R
\name{add_audio_analysis}
\alias{add_audio_analysis}
\title{Add Spotify audio analysis to a data frame}
\usage{
add_audio_analysis(df, ...)
}
\arguments{
\item{df}{A data frame with a \code{track_uri} column.}

\item{...}{Parameters passed on to \code{\link{get_tidy_audio_analysis}}.}
}
\description{
Fetches and joins the Spotify audio analysis for every Spotify URI in a data
frame.
}
\examples{
library(tidyverse)
spotifyr::get_playlist_audio_features("", "37i9dQZF1DX21bRPJuEN7r") \%>\%
  slice(1:5) \%>\%
  add_audio_analysis()
}
\seealso{
\code{\link{get_tidy_audio_analysis}}
}
