% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{compmus_align}
\alias{compmus_align}
\title{Aligns lower-level Spotify segmentations with higher-level segmentations}
\usage{
compmus_align(dat, outer, inner)
}
\arguments{
\item{dat}{A tibble.}

\item{outer, inner}{Tibble columns with \code{start} and \code{duration}
columns.}
}
\description{
Embeds the \code{inner} segments as a list column with sub-tibbles
corresponding to each \code{outer} segment.
}
\examples{
library(tidyverse)
get_tidy_audio_analysis('5ZLkc5RY1NM4FtGWEd6HOE') \%>\%
    compmus_align(bars, segments) \%>\%
    select(bars) \%>\% unnest(bars)
}
