% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwrapping.R
\name{compmus_c_transpose}
\alias{compmus_c_transpose}
\title{Transpose chroma vectors to C}
\usage{
compmus_c_transpose(dat, key)
}
\arguments{
\item{dat}{Data frame with a \code{pitches} column containing chroma
vectors.}

\item{key}{Character string naming the global or local tonal centre.}
}
\description{
Given a key estimate, transpose chroma vectors back to C.
}
\examples{
library(tidyverse)
spotifyr::get_playlist_audio_features("", "37i9dQZF1DWWEcRhUVtL8n") \%>\%
  slice(1:20) \%>\%
  add_audio_analysis() \%>\%
  mutate(segments = map2(segments, key, compmus_c_transpose))
}
