% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwrapping.R
\name{compmus_gather_chroma}
\alias{compmus_gather_chroma}
\alias{compmus_gather_timbre}
\title{Gather chroma vectors}
\usage{
compmus_gather_chroma(dat)

compmus_gather_timbre(dat)
}
\arguments{
\item{dat}{Data frame with a \code{pitches} or \code{timbre} column.}
}
\description{
These functions gather chroma or timbre vectors into long format for
Computational Musicology, as generated by
\code{\link{get_tidy_audio_analysis}}.
}
\section{Functions}{
\itemize{
\item \code{compmus_gather_timbre}: Gather timbre vectors
}}

\examples{
library(tidyverse)
get_tidy_audio_analysis('6IQILcYkN2S2eSu5IHoPEH') \%>\%
    select(segments) \%>\% unnest(segments) \%>\%
    compmus_gather_chroma

get_tidy_audio_analysis('5ZLkc5RY1NM4FtGWEd6HOE') \%>\%
    select(segments) \%>\% unnest(segments) \%>\%
    compmus_gather_timbre
}
