% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norms.R
\name{compmus_match_pitch_template}
\alias{compmus_match_pitch_template}
\title{Match chroma vectors against templates}
\usage{
compmus_match_pitch_template(
  dat,
  templates,
  method = "cosine",
  norm = "euclidean"
)
}
\arguments{
\item{dat}{A data frame containing chroma vectors in a \code{pitches} column.}

\item{templates}{A data frame with a \code{name} column for each template and
the templates themselves in a \code{template} column.}

\item{method}{A character string indicating which distance metric to use (see
\code{\link{compmus_long_distance}}). Default is cosine distance.}

\item{norm}{An optional character string indicating the method for
pre-normalising each vector with \code{\link{compmus_normalise}}. Default
is Euclidean.}
}
\value{
A tibble with columns \code{start}, \code{duration}, \code{name}, and
  \code{d}.
}
\description{
Compares chroma vectors in a data frame against a list of templates, most
likely key or chord profiles.
}
