% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norms.R
\name{compmus_normalise}
\alias{compmus_normalise}
\title{Normalise vectors}
\usage{
compmus_normalise(v, method = "euclidean")
}
\arguments{
\item{v}{A numeric vector.}

\item{method}{A character string indicating which normalization to use (see
Details). Default is the Euclidean norm.}
}
\description{
We use a number of vector normalisation strategies in Computational
Musicology. This function brings them together into one place, along with
common alternative names.
}
\details{
The following methods are supported.
\describe{
    \item{\code{identity},\code{id}}{No normalisation.}
    \item{\code{harmonic}}{Harmonic mean.}
    \item{\code{manhattan},\code{L1}}{Manhattan (L1) norm.}
    \item{\code{euclidean},\code{L2}}{Euclidean (L2) norm.}
    \item{\code{chebyshev},\code{maximum}}{Chebyshev (maximum) norm.}
    \item{\code{aitchison},\code{clr}}{Aitchison's clr transformation.}
    \item{\code{softmax}}{Softmax.}}
}
\examples{
library(tidyverse)
get_tidy_audio_analysis('6IQILcYkN2S2eSu5IHoPEH') \%>\%
    select(segments) \%>\% unnest(segments) \%>\%
    mutate(pitches = map(pitches, compmus_normalise, 'euclidean'))
}
