% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{compmus_summarise}
\alias{compmus_summarise}
\alias{compmus_summarize}
\title{Summarise possibly vector-valued columns}
\usage{
compmus_summarise(dat, feature, method = "mean", norm = "id",
  na.rm = FALSE)

compmus_summarize(dat, feature, method = "mean", norm = "id",
  na.rm = FALSE)
}
\arguments{
\item{dat}{A tibble containing list columns.}

\item{feature}{The vector-valued column of \code{dat} to summarise (unquoted).}

\item{method}{A character string indicating which summarisation method to use
(see Details). Default is the arithmetic mean.}

\item{norm}{An optional character string indicating the method for
pre-normalising each vector with \code{\link{compmus_normalise}}.}

\item{na.rm}{Remove NA values?}
}
\description{
Summarises vector-based features in nested list columns.
}
\details{
The following methods are supported.
}
\section{Functions}{
\itemize{
\item \code{compmus_summarize}: Summarize possibly vector-valued columns
}}

\section{Central Tendency}{


\describe{
    \item{\code{mean}}{arithmetic mean}
    \item{\code{aitchison},\code{acentre},\code{acenter}}{Aitchison centre}
    \item{\code{rms}}{root mean square}
    \item{\code{max}}{maximum}}
}

\section{Dispersion}{


\describe{
    \item{\code{sd}}{standard deviation}
    \item{\code{asd}}{standard deviation of clr-transformed components}
    \item{\code{sdsq}}{standard deviation of squares}
    \item{\code{varration}}{variation ratio}}
}

\examples{
library(tidyverse)
get_tidy_audio_analysis('5ZLkc5RY1NM4FtGWEd6HOE') \%>\%
    compmus_align(bars, segments) \%>\%
    select(bars) \%>\% unnest(bars) \%>\%
    mutate(
        pitches =
            map(
                segments,
                compmus_summarise, pitches,
                method = 'rms', norm = 'euclidean'))
}
