% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_analysis.R
\name{get_tidy_audio_analysis}
\alias{get_tidy_audio_analysis}
\title{Get Spotify audio analysis tidily}
\usage{
get_tidy_audio_analysis(track_uri, ...)
}
\arguments{
\item{track_uri}{A string with a Spotify track URI.}

\item{...}{Additional parameters passed to
\code{\link[spotifyr]{get_track_audio_analysis}}.}
}
\description{
Fetches the Spotify audio analysis for a track using list columns rather than
lists of lists.
}
\details{
See the
\href{https://developer.spotify.com/documentation/web-api/reference/tracks/get-audio-analysis/}{Spotify
developer documentation} for details about the information included in an
audio analysis.
}
\examples{
get_tidy_audio_analysis("6IQILcYkN2S2eSu5IHoPEH")
}
