% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwrapping.R
\name{compmus_gather_chroma}
\alias{compmus_gather_chroma}
\alias{compmus_gather_timbre}
\title{Gather chroma/timbre vectors}
\usage{
compmus_gather_chroma(dat)

compmus_gather_timbre(dat)
}
\arguments{
\item{dat}{Data frame with a \code{pitches} or \code{timbre} column.}
}
\description{
These functions gather chroma or timbre vectors into long format.
}
\section{Functions}{
\itemize{
\item \code{compmus_gather_timbre()}: Gather chroma/timbre vectors

}}
