% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norms.R
\name{compmus_match_pitch_template}
\alias{compmus_match_pitch_template}
\title{Match chroma vectors against templates}
\usage{
compmus_match_pitch_template(dat, templates, method = "cosine",
  norm = "euclidean")
}
\arguments{
\item{dat}{A data frame containing chroma vectors in a \code{pitches} column.}

\item{templates}{A data frame with a \code{name} column for each template and
the templates themselves in a \code{template} column.}

\item{method}{A character string indicating which distance metric to use (see
\code{\link{compmus_long_distance}}). Default is cosine distance.}

\item{norm}{An optional character string indicating the method for
pre-normalising each vector with \code{\link{compmus_normalise}}. Default
is Euclidean.}
}
\value{
A tibble with columns \code{start}, \code{duration}, \code{name}, and
  \code{d}.
}
\description{
Compares chroma vectors in a data frame against a list of templates, most
likely key or chord profiles.
}
\examples{
library(tidyverse)
circshift <- function(v, n) {if (n == 0) v else c(tail(v, n), head(v, -n))}
major_chord <-
    c(1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0)
minor_chord <-
    c(1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0)
chord_templates <-
    tribble(
        ~name   , ~template,
        'D:min' , circshift(minor_chord,  2),
        'F:maj' , circshift(major_chord,  5),
        'A:min' , circshift(minor_chord,  9),
        'C:maj' , circshift(major_chord,  0),
        'E:min' , circshift(minor_chord,  4),
        'G:maj' , circshift(major_chord,  7),
        'B:min' , circshift(minor_chord, 11))

get_tidy_audio_analysis('5UVsbUV0Kh033cqsZ5sLQi') \%>\%
    compmus_align(sections, segments) \%>\%
    select(sections) \%>\% unnest(sections) \%>\%
    mutate(
        pitches =
            map(segments,
                compmus_summarise, pitches,
                method = 'mean', norm = 'manhattan')) \%>\%
    compmus_match_pitch_template(chord_templates, 'euclidean', 'manhattan')
}
